+++ web/src/main/java/com/garret/movies/config/mapper/ModelMapperConfigurer.java
@@ -1,0 +1,5 @@
+package com.garret.movies.config.mapper;
+
+public interface ModelMapperConfigurer {
+    void configure();
+}
